Support Markdown files in docs
To make it easier for people to update the docs without learning
reStructuredText, support rendering of Markdown using recommonmark.

See http://blog.readthedocs.com/adding-markdown-support/ for
details.

I needed to upgrade Sphinx to a newer version for the updated
configuration to work.

You will need to re-run:

    workon docs
    pip install -r requirements.txt

to get this to work.

The source files still need to be converted to Markdown, and I'm
not sure if it's a good idea to convert all files to Markdown.

More on this to come.

Addresses #1111
https://tribune.unfuddle.com/a#/projects/6/tickets/by_number/1111