Fix troubleshooting structure and links
We're still learning the best ways to mix Markdown and rST.

Include the sub-pages in `tarbell` in a `toctree` so we can link to
them more easily.  Be sure to include the new `tarbell/troubelshooting`
there as well.

Use rST syntax when linking to a section in the same document inside a
note.