Sectionize Tarbell troubleshooting doc
Geoff thinks this is clearer, both visually and semantically.

Also, it's less markup and indentation to replicate as other
contributors add sections.

Also, as people copy and paste to add entries, it gets them in the habit
of thinking in terms of semantic document structure rather than visual
separation of information.